Add initial version of example generator (#24)
Adds a new extension that can generate example projects into the current workspace.
As an initial starting point, it adds a plain CMake project example with launch.json,
the example project of clangd-contexts, and traces for TraceCompass.cloud.

The example generation is triggered via a command. In the arguments of this command
you can optionally already pre-specify the example type.

This extension is extensible in the sense that you can add more examples by adding a
new example folder and extending the command asking you which example to generate.

Also
* Fix getting started page in browser app
* Fix browser app and add default workspace

Fixes #12

Signed-off-by: Philip Langer <planger@eclipsesource.com>
Co-authored-by: Simon Graband <sgraband@eclipsesource.com>
Co-authored-by: Stefan Dirix <sdirix@eclipsesource.com>

Contributed on behalf of STMicroelectronics.