fix: windows deployment
The Jenkinsfile expects the version number of the previous release(s)
in the Windows script, not the current one. This version number is
used to indicate which of the old Windows versions can be updated to
the current one. This is now fixed, re-enabling the Windows update
mechanism.

Contributed on behalf of STMicroelectronics